Move formio api routes to a new formio2022.js file, for easier maintainability with new 2023 formio routes
