+++ docs/csb-openapi.json
@@ -492,9 +492,9 @@
         ]
       }
     },
-    "/api/formio-2022-frf-submissions": {
+    "/api/formio/2022/frf-submissions": {
       "get": {
-        "summary": "/api/formio-2022-frf-submissions",
+        "summary": "/api/formio/2022/frf-submissions",
         "responses": {
           "200": {
             "description": "OK",
@@ -514,9 +514,9 @@
         "parameters": [{ "$ref": "#/components/parameters/scan" }]
       }
     },
-    "/api/formio-2022-frf-submission": {
+    "/api/formio/2022/frf-submission": {
       "post": {
-        "summary": "/api/formio-2022-frf-submission",
+        "summary": "/api/formio/2022/frf-submission",
         "responses": {
           "200": {
             "description": "OK",
@@ -536,9 +536,9 @@
         "parameters": [{ "$ref": "#/components/parameters/scan" }]
       }
     },
-    "/api/formio-2022-frf-submission/{mongoId}": {
+    "/api/formio/2022/frf-submission/{mongoId}": {
       "get": {
-        "summary": "/api/formio-2022-frf-submission/{mongoId}",
+        "summary": "/api/formio/2022/frf-submission/{mongoId}",
         "parameters": [
           {
             "name": "mongoId",
@@ -586,7 +586,7 @@
         "tags": []
       },
       "post": {
-        "summary": "/api/formio-2022-frf-submission/{mongoId}",
+        "summary": "/api/formio/2022/frf-submission/{mongoId}",
         "parameters": [
           {
             "name": "mongoId",
@@ -613,9 +613,9 @@
         "tags": []
       }
     },
-    "/api/formio-2022-prf-submissions": {
+    "/api/formio/2022/prf-submissions": {
       "get": {
-        "summary": "/api/formio-2022-prf-submissions",
+        "summary": "/api/formio/2022/prf-submissions",
         "responses": {
           "200": {
             "description": "OK",
@@ -635,9 +635,9 @@
         "parameters": [{ "$ref": "#/components/parameters/scan" }]
       }
     },
-    "/api/formio-2022-prf-submission": {
+    "/api/formio/2022/prf-submission": {
       "post": {
-        "summary": "/api/formio-2022-prf-submission",
+        "summary": "/api/formio/2022/prf-submission",
         "responses": {
           "200": {
             "description": "OK",
@@ -657,9 +657,9 @@
         "parameters": [{ "$ref": "#/components/parameters/scan" }]
       }
     },
-    "/api/formio-2022-prf-submission/{rebateId}": {
+    "/api/formio/2022/prf-submission/{rebateId}": {
       "get": {
-        "summary": "/api/formio-2022-prf-submission/{rebateId}",
+        "summary": "/api/formio/2022/prf-submission/{rebateId}",
         "parameters": [
           {
             "name": "rebateId",
@@ -707,7 +707,7 @@
         "tags": []
       },
       "post": {
-        "summary": "/api/formio-2022-prf-submission/{rebateId}",
+        "summary": "/api/formio/2022/prf-submission/{rebateId}",
         "parameters": [
           {
             "name": "rebateId",
@@ -734,9 +734,9 @@
         "tags": []
       }
     },
-    "/api/delete-formio-2022-prf-submission": {
+    "/api/formio/2022/delete-prf-submission": {
       "post": {
-        "summary": "/api/delete-formio-2022-prf-submission",
+        "summary": "/api/formio/2022/delete-prf-submission",
         "responses": {
           "200": {
             "description": "OK",
@@ -753,9 +753,9 @@
         "parameters": [{ "$ref": "#/components/parameters/scan" }]
       }
     },
-    "/api/formio-2022-crf-submissions": {
+    "/api/formio/2022/crf-submissions": {
       "get": {
-        "summary": "/api/formio-2022-crf-submissions",
+        "summary": "/api/formio/2022/crf-submissions",
         "responses": {
           "200": {
             "description": "OK",
@@ -775,9 +775,9 @@
         "parameters": [{ "$ref": "#/components/parameters/scan" }]
       }
     },
-    "/api/formio-2022-crf-submission": {
+    "/api/formio/2022/crf-submission": {
       "post": {
-        "summary": "/api/formio-2022-crf-submission",
+        "summary": "/api/formio/2022/crf-submission",
         "responses": {
           "200": {
             "description": "OK",
@@ -797,9 +797,9 @@
         "parameters": [{ "$ref": "#/components/parameters/scan" }]
       }
     },
-    "/api/formio-2022-crf-submission/{rebateId}": {
+    "/api/formio/2022/crf-submission/{rebateId}": {
       "get": {
-        "summary": "/api/formio-2022-crf-submission/{rebateId}",
+        "summary": "/api/formio/2022/crf-submission/{rebateId}",
         "parameters": [
           {
             "name": "rebateId",
@@ -847,7 +847,7 @@
         "tags": []
       },
       "post": {
-        "summary": "/api/formio-2022-crf-submission/{rebateId}",
+        "summary": "/api/formio/2022/crf-submission/{rebateId}",
         "parameters": [
           {
             "name": "rebateId",
